chore(backend): move utils to qnaasm
