+++ src/qnaasm/qfree.py
@@ -1,8 +1,7 @@
 from typing import TYPE_CHECKING
 
 from qnaasm.qnaasm import QNAasm
-
-from utils.position import Position
+from qnaasm.position import Position
 
 if TYPE_CHECKING:
     from qnaasm.visitor import Visitor
