+++ src/qnaasm_main.py
@@ -10,9 +10,8 @@ from qnaasm.nodes import (
     QNAasm,
 )
 
-from utils.classical_register import ClassicalRegister
-from utils.position import Position
-
+from qnaasm.classical_register import ClassicalRegister
+from qnaasm.position import Position
 from qnaasm.pretty_printer import PrettyPrinter
